decomp(mispair-form): OutfitConfig::MatSwap::Compress + CharCuff::Highlight → 100%
MatSwap::Compress: signed cast (int)mat forces cmpwi over cmplwi.
CharCuff::Highlight: explicit-parens multiply-chain order forces unfolded
2-fmuls codegen matching sibling line. +2 strict (18874→18876).

+++ src/system/bandobj/OutfitConfig.cpp
@@ -202,7 +202,7 @@ void OutfitConfig::MatSwap::Compose(
 
 bool OutfitConfig::MatSwap::Compress(BandCharDesc *desc) {
     RndMat *mat = mMat;
-    if (!mat)
+    if (!(int)mat)
         return false;
     RndTex *diffTex = mat->GetDiffuseTex();
     if (!diffTex || (diffTex->GetType() & RndTex::kRenderedNoZ) != RndTex::kRenderedNoZ)
